pr review corrections
