+++ src/App.tsx
@@ -3,12 +3,12 @@ import React from 'react';
 import {NavigationContainer} from '@react-navigation/native';
 import Routes from './navigation';
 
-function App() {
+const App: React.FC = () => {
     return (
         <NavigationContainer>
             <Routes />
         </NavigationContainer>
     );
-}
+};
 
 export default App;
